refactor: create util directory for common functionality
The following modules/functions were moved:
* ``match.py`` moved from the add module
* ``fmt_album_changes`` moved from ``cli.py``
* ``PromptChoice`` moved from ``cli.py``
* ``choice_prompt`` moved from ``cli.py``
