--- moe/plugins/add/add_core/__init__.py
@@ -1,1 +1,0 @@
-"""Core api add plugin."""
