+++ moe/plugins/musicbrainz/mb_cli.py
@@ -16,17 +16,14 @@ from moe.config import Config
 from moe.library.album import Album
 from moe.plugins import moe_import
 from moe.plugins import musicbrainz as moe_mb
+from moe.util.cli import PromptChoice
 
 
 @moe.hookimpl
-def add_import_prompt_choice(prompt_choices: List[moe.cli.PromptChoice]):
+def add_import_prompt_choice(prompt_choices: List[PromptChoice]):
     """Adds the ``apply`` and ``abort`` prompt choices to the user prompt."""
     prompt_choices.append(
-        moe.cli.PromptChoice(
-            title="Enter Musicbrainz ID",
-            shortcut_key="m",
-            func=_enter_id,
-        )
+        PromptChoice(title="Enter Musicbrainz ID", shortcut_key="m", func=_enter_id)
     )
