+++ moe/util/core/__init__.py
@@ -1,0 +1,7 @@
+"""This package contains shared functionality for the core API."""
+
+from . import match
+from .match import *
+
+__all__ = []
+__all__.extend(match.__all__)
