+++ tests/plugins/add/test_add_cli.py
@@ -12,6 +12,7 @@ import moe.cli
 from moe.config import Config
 from moe.library.album import Album
 from moe.plugins import add
+from moe.util.cli import PromptChoice
 
 
 @pytest.fixture
@@ -91,7 +92,7 @@ class TestPreAdd:
         """When adding a track, call the function with the respective album."""
         tmp_session.merge(mock_track)
         db_album = mock_track.album_obj.get_existing()
-        with patch.object(moe.cli, "choice_prompt") as mock_prompt_choice:
+        with patch("moe.plugins.add.add_cli.choice_prompt") as mock_prompt_choice:
             tmp_add_config.plugin_manager.hook.pre_add(
                 config=tmp_add_config, item=mock_track
             )
@@ -104,7 +105,7 @@ class TestPreAdd:
         """When adding an extra, call the function with the respective album."""
         tmp_session.merge(mock_extra)
         db_album = mock_extra.album_obj.get_existing()
-        with patch.object(moe.cli, "choice_prompt") as mock_prompt_choice:
+        with patch("moe.plugins.add.add_cli.choice_prompt") as mock_prompt_choice:
             tmp_add_config.plugin_manager.hook.pre_add(
                 config=tmp_add_config, item=mock_extra
             )
@@ -126,8 +127,10 @@ class TestPreAdd:
         dup_album = tmp_session.merge(dup_album)
         assert album.get_existing()
 
-        mock_prompt_choice = moe.cli.PromptChoice("mock", "m", add.add_cli._replace)
-        with patch.object(moe.cli, "choice_prompt", return_value=mock_prompt_choice):
+        mock_prompt_choice = PromptChoice("mock", "m", add.add_cli._replace)
+        with patch(
+            "moe.plugins.add.add_cli.choice_prompt", return_value=mock_prompt_choice
+        ):
             tmp_add_config.plugin_manager.hook.pre_add(
                 config=tmp_add_config, item=album
             )
@@ -149,8 +152,10 @@ class TestPreAdd:
         dup_album = tmp_session.merge(dup_album)
         assert album.get_existing()
 
-        mock_prompt_choice = moe.cli.PromptChoice("mock", "m", add.add_cli._abort)
-        with patch.object(moe.cli, "choice_prompt", return_value=mock_prompt_choice):
+        mock_prompt_choice = PromptChoice("mock", "m", add.add_cli._abort)
+        with patch(
+            "moe.plugins.add.add_cli.choice_prompt", return_value=mock_prompt_choice
+        ):
             with pytest.raises(SystemExit) as error:
                 tmp_add_config.plugin_manager.hook.pre_add(
                     config=tmp_add_config, item=album
@@ -172,8 +177,10 @@ class TestPreAdd:
         tmp_session.merge(dup_album)
         assert new_album.get_existing()
 
-        mock_prompt_choice = moe.cli.PromptChoice("mock", "m", add.add_cli._merge)
-        with patch.object(moe.cli, "choice_prompt", return_value=mock_prompt_choice):
+        mock_prompt_choice = PromptChoice("mock", "m", add.add_cli._merge)
+        with patch(
+            "moe.plugins.add.add_cli.choice_prompt", return_value=mock_prompt_choice
+        ):
             tmp_add_config.plugin_manager.hook.pre_add(
                 config=tmp_add_config, item=new_album
             )
@@ -198,8 +205,10 @@ class TestPreAdd:
         tmp_session.merge(dup_album)
         assert new_album.get_existing()
 
-        mock_prompt_choice = moe.cli.PromptChoice("mock", "m", add.add_cli._overwrite)
-        with patch.object(moe.cli, "choice_prompt", return_value=mock_prompt_choice):
+        mock_prompt_choice = PromptChoice("mock", "m", add.add_cli._overwrite)
+        with patch(
+            "moe.plugins.add.add_cli.choice_prompt", return_value=mock_prompt_choice
+        ):
             tmp_add_config.plugin_manager.hook.pre_add(
                 config=tmp_add_config, item=new_album
             )
@@ -212,7 +221,7 @@ class TestPreAdd:
 
     def test_no_dup_album(self, tmp_session, mock_album, tmp_add_config):
         """Don't do anything if the album does not have a duplicate."""
-        with patch.object(moe.cli, "choice_prompt") as mock_choice_prompt:
+        with patch("moe.plugins.add.add_cli.choice_prompt") as mock_choice_prompt:
             tmp_add_config.plugin_manager.hook.pre_add(
                 config=tmp_add_config, item=mock_album
             )
@@ -221,7 +230,7 @@ class TestPreAdd:
 
     def test_no_dup_track(self, tmp_session, mock_track, tmp_add_config):
         """Don't do anything if the track does not have a duplicate."""
-        with patch.object(moe.cli, "choice_prompt") as mock_choice_prompt:
+        with patch("moe.plugins.add.add_cli.choice_prompt") as mock_choice_prompt:
             tmp_add_config.plugin_manager.hook.pre_add(
                 config=tmp_add_config, item=mock_track
             )
@@ -230,7 +239,7 @@ class TestPreAdd:
 
     def test_no_dup_extra(self, tmp_session, mock_extra, tmp_add_config):
         """Don't do anything if the extra does not have a duplicate."""
-        with patch.object(moe.cli, "choice_prompt") as mock_choice_prompt:
+        with patch("moe.plugins.add.add_cli.choice_prompt") as mock_choice_prompt:
             tmp_add_config.plugin_manager.hook.pre_add(
                 config=tmp_add_config, item=mock_extra
             )
